ref: extract library grouping logic into isolated UseCase
💡 What:
Extracted `groupByDynamic`, `groupByUngrouped`, and `groupByCategory` functions from `LibraryViewModel.kt` into a new `GroupLibraryMangaUseCase.kt` file.

🎯 Why:
To reduce the size of the massive `LibraryViewModel`, improve separation of concerns, and isolate the logic that determines how library manga items are grouped.

📍 From/To:
Moved logic from `app/src/main/java/eu/kanade/tachiyomi/ui/library/LibraryViewModel.kt` to the newly created `app/src/main/java/org/nekomanga/usecases/library/GroupLibraryMangaUseCase.kt`.
