+++ .jules/renovator.md
@@ -1,2 +1,3 @@
 ## 2026-03-24 - [Avoid InjektionException when extracting UseCases] **Learning:** [Kohesive Injekt is a service locator that requires explicit registration in a DI module (e.g., an `AppModule`). Injecting new unregistered UseCases via `private val mergeMangaUseCases: MergeMangaUseCases = Injekt.get()` causes runtime crashes.] **Action:** [When creating new UseCases, instantiate them directly (e.g., `private val mergeMangaUseCases = MergeMangaUseCases()`) and rely on the default `Injekt.get()` parameters inside their constructors to resolve framework dependencies.]
 ## 2026-03-24 - [Extracting Category Logic to UseCases] **Learning:** [When extracting common functionality like Category management out of ViewModels into an isolated UseCase, it is essential to trace all ViewModels utilizing that logic (e.g., `LibraryViewModel` and `MangaViewModel`) to ensure uniform extraction. Additionally, if the UseCase is not registered in the DI module (`AppModule`), it must be instantiated directly inside the ViewModel or another UseCase class wrapper, relying on default constructor parameters for dependency injection, to avoid runtime `InjektionException` crashes.] **Action:** [Trace usage of domain concepts across all ViewModels before extraction, and verify DI registration status for newly created UseCase boundaries to prevent runtime injection failures.]
+## 2026-04-24 - [Extract ViewModel logic to UseCase] **Learning:** [When extracting complex logic (like grouping lists dynamically) from a large ViewModel into a separate UseCase, it's safer to use simple extension functions or directly call instantiated class methods instead of injecting the UseCase via `Injekt.get()` if it hasn't been registered in the DI graph. This prevents `InjektionException` crashes.] **Action:** [Instantiate simple pure-function UseCases directly (e.g., `val groupLibraryManga = GroupLibraryMangaUseCase()`) inside the ViewModels to maintain isolation without touching the DI graph unless necessary.]
